Complete implementation of enrichOffencesFromCourtCasesAndMatcherOutcome and tweak structure
