+++ src/types/ExceptionCode.ts
@@ -19,6 +19,7 @@ export enum ExceptionCode {
   HO100220 = "HO100220",
   HO100304 = "HO100304",
   HO100310 = "HO100310",
+  HO100320 = "HO100320",
   HO100322 = "HO100322",
   HO100328 = "HO100328",
   HO100329 = "HO100329",
